+++ podcasts/MainTabBarController.swift
@@ -28,6 +28,26 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
     /// Displayed during database migrations
     var alert: ShiftyLoadingAlert?
 
+    func loginAgain() {
+        // Ensure the new sync is a full sync (so podcasts and episodes are retrieved)
+        SyncManager.syncReason = .login
+        ServerSettings.clearLastSyncTime()
+        UserDefaults.standard.removeObject(forKey: "PCLastModifiedServerDate")
+
+        // Copy data from the previous corrupted database (if possible)
+        alert = ShiftyLoadingAlert(title: "Corrupted database. Recovering...")
+        alert?.showAlert(self, hasProgress: false, completion: nil)
+        DataManager.sharedManager.copyAllData()
+
+        alert?.hideAlert(true, completion: {
+            // Start the full sync
+            let controller = SyncSigninViewController()
+            controller.loginAgain = true
+            SceneHelper.rootViewController()?.dismiss(animated: true)
+            SceneHelper.rootViewController()?.present(controller, animated: true, completion: nil)
+        })
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -105,6 +125,10 @@ class MainTabBarController: UITabBarController, NavigationProtocol {
 
         updateDatabaseIndexes()
         optimizeDatabaseIfNeeded()
+
+        if DataManager.loginAgain {
+            loginAgain()
+        }
     }
 
     /// Update database indexes and delete unused columns